# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "e-Demic API Documentation",
        "description": "API endpoints and descriptions for e-Demic Laravel API",
        "contact": {
            "name": "Developer : Sreejith B S",
            "email": "sreejith.bs@quinoid.in"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/messages/{index}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "List all messages",
                "description": "API to list all messages, such as alerts, announcements etc on latest first basis. Default limit set as **10** messages per page. <br><br>Accepts `index` for lazy load messages.",
                "operationId": "messages",
                "parameters": [
                    {
                        "name": "index",
                        "in": "path",
                        "description": "Page index to fetch messages",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "/diseases": {
            "get": {
                "tags": [
                    "Diseases"
                ],
                "summary": "List all diseases",
                "description": "API to list all diseases.",
                "operationId": "diseases",
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "/patients": {
            "get": {
                "tags": [
                    "Map"
                ],
                "summary": "List all patient details for map plotting",
                "description": "API to list all patient details for map plotting.",
                "operationId": "diseases",
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}